requirements.txt
# requirements for Pokemon fetcher
requests>=2.31.0
psycopg2-binary>=2.9.9

# API dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
